[feat] duration  int -> float 변경 - #26

+++ dateroad-domain/src/main/java/org/dateroad/place/domain/DatePlace.java
@@ -23,7 +23,7 @@ public class DatePlace extends Place {
     @NotNull
     private Date date;
 
-    public static DatePlace create(final String name, int duration, final Date date, int sequence) {
+    public static DatePlace create(final String name, float duration, final Date date, int sequence) {
         return DatePlace.builder()
                 .name(name)
                 .duration(duration)
